Refactored action types
